Equipment management and lanterns

+++ Inventory Scripts/ItemTypeScripts/Equipment.cs
@@ -9,7 +9,7 @@ public class Equipment : Item
     public EquipmentSlot equipSlot;
     public int armorModifier;
     public int damageModifer;
-    public int speedModifer;
+    public float speedModifer;
     public override void Use()
     {
         base.Use();
